Refactor open API response generator
We had 2 different functions for creating open API responses. Those functions had almost the same logic. The only difference is that the first function worked only with an object as an input and the second function worked with an array as an input, but the api responses creation logic was still the same. It's a bad pattern to duplicate logic in different places.

 There are 2 different solutions for that:
 1) Make another abstraction layer that has the logic that is used in both functions;
 2) Refactor functions.

 I decided to go with the second approach since the first one would make everything more difficult in my opinion.
 So, the commit merges two functions into one which accepts both, an object and an array as an input. So, the client doesn't need to care about the function implementation or think about which function should I use. Now the client just uses one function in any places.