Add logic for getting signature on frontend
Despite the fact that I don't need to implement the frontend part in the tech task I still want to provide a simple UI for getting a signature for the user.
The signature is used in backend testing on Swagger, so it's better to have the logic for getting this signature already. It makes the review process the way easier.