Extend getting balances logic
Before `GET /balances/:address` returned just a list of balances for the given account. I recognised that Alchemy API also returns the page key for the next page if the user has more than X amount of balances per page. So, the commit extends this logic by including `pageKey` in the route. Now, the route supports different pages and the client can get all balances for the account.