Introduce leaderboard route
The commit adds logic for the leaderboard route. It allows the client getting the leaderboard list of users which are sorted by the sum up of all tokens they own.
It looks like the route is quite simple, but the commit contains a lot of logic. Here is the breakdown:
1) Adds a postgres database. For that 2 libraries are installed `prisma` and `@prisma/client`. They make the interaction with postgres really convenient. The database itself is running via docker;
2) A table `User` was created in this database. It contains data of all users that have logged in the application. Based on this data the leaderboard list is created;
3) New route `GET /user/leaderboard` was introduced;
4) During logging process the user is inserted in the database if they aren't there yet;
5) Swagger documentation is updated with last changes;
6) All new functionality is covered with tests.