Fix bug with not-updated balance values in leaderboard
There was a silly mistake by me in the leaderboard logic. I created the user's record in the database only during the logging process and I haven't updated it which is completely wrong. The commit fixes this bug by updating the user's balance every time the user logs in the application.