Move env variables into docker
The commit adds env variables directly to the docker service to make testing the application easier. Now, the person who's going to test the project can start the project just with one command.
I know that leaking private information from env variable is wrong, but for the test project I decided to keep it in this way on purpose to make setup easier.